Corrected version:
This is the first commit of this project.
First, I tried to add CMSIS-DSP files, and then I will write some
functions to generate signals and implement a low-pass filter.

I need to analyze filter external samples, so I created some functions
to send samples to MATLAB and display them in plots.